requirements.txt
# Core AI / model dependencies
torch>=2.1.0
torchvision>=0.16.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0

# Web interface
streamlit>=1.30.0

# Utilities
requests>=2.31.0
pydantic>=2.1.0
tqdm>=4.66.5

# Optional: for debugging / image display
matplotlib>=3.8.0
scipy>=1.12.0
